+++ src/extension.ts
@@ -3,7 +3,7 @@ import { telemetry, initializeTelemetry } from "./telemetry";
 import { CreateArchitectureCommand } from "./commands/create-architecture";
 
 export function activate(context: vscode.ExtensionContext): void {
-    initializeTelemetry(context);
+    initializeTelemetry();
     registerCommand(context);
     telemetry.sendTelemetryEvent("extensionActivated");
 }
@@ -16,4 +16,7 @@ function registerCommand(context: vscode.ExtensionContext): void {
     context.subscriptions.push(disposable);
 }
 
-export function deactivate(): void {}
+export async function deactivate(): Promise<void> {
+    telemetry.sendTelemetryEvent("extensionDeactivated");
+    await telemetry.dispose();
+}